+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `check` command: validate content directory without building (`simple-gal check`)
+
 ## [0.4.1] - 2026-02-16
 
 ### Fixed